zram: add zcomp asynchronous function utilities
To support asynchrnous compression, zcomp_cookie need to be alive
until compresion is done by backend so we should allocate it on
heap. However, the allocation/free overhead is not trivial when
the backend is really fast.

This patch introduces zcomp_cookie_pool to avoid the overhead
for asynchronous compression backend.

Bug: 167626384
Signed-off-by: Minchan Kim <minchan@google.com>
Change-Id: I1942b6124dadb97303a52fe4c7eff0a41122a421