arm64/dts: slider: Use AP->GSA->GSC interface
Slider can be configured to talk to Dauntless (GSC) directly (for
early bringup), or through GSA (like Whitefin V2 and all later
devices).

This causes the kernel to use the GSA interface. You'll still
need to flip some switches to connect the SPI bus correctly.

  SW17051 -> ON
  SW23041 -> ON

Early sliders only have one of these switches (with a different
label). There's a photo showing the switch locations in the bug.

Note that the Slider bootloader has already made this change. If
"fastboot oem gsc version" is not able to communicate with
Dauntless, try flipping the switches.

Bug: b/168136227
Test: manual
Signed-off-by: Bill Richardson <wfrichar@google.com>
Change-Id: Ieff44622eedfba8033dd0202b0032c00e539d438