bts: Use BUS1 to INT mapping table to vote INT freq
Currently in the BTS driver we vote the INT freq based on the
following equation:

  btsdev->peak_bw / BUS_WIDTH * 100 / INT_UTIL;

However, the calculated freq should be the clk freq of BUS1 so we
should use a mapping table to get the final INT freq. This would
reduce the INT freq vote and could result in lower power.

Bug: 177367696
Test: Boot & check BTS INT votes
Signed-off-by: Jia-yi Chen <jychen@google.com>
Change-Id: If5a4031e7f7bc8e4ed70df9ae51ead6dee9bbfdb