usb: typec: tcpci_fusb307b: Enable data path either in BC1.2 or PD path
USB comm capable notification arrives after the partner is determined
as pd capable or not. This causes a temporary disconnect.
Prevent this by leaving the data path enabled if BC1.2 indicates as
SDP or CDP port.

Similar to 89d0887614fb9ba3206885b49d5548d746cd2cf1 but on fusb307b.

Bug: 178839617
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Change-Id: I7792846dc05cbbace9e7366afe0737b3c807b4a1