vendor_hook: sched: Add a filenode to control task spreading
Currently, if the uclamp.min of a task is over the capacity of a
cpu, it will be scheduled on that cpu. Add a filenode to control
this mechanism, if enabled, the scheduler will not check the
uclamp.min of that task, so that tasks will be more spreading on
the cpus. Also get rid of the capacity margin checking for boosted
tasks.

Bug: 170510392
Bug: 170508815
Signed-off-by: Rick Yiu <rickyiu@google.com>
Change-Id: I5f5c678074fd0082d9caba1ca04898f69cba07f5