vendor_hook: sched: Let scheduler skip util checking if cpu is idle
Current cpu util includes util of runnable tasks plus the recent
utilization of currently non-runnable tasks, so it may return a non-zero
value even there is no task running on a cpu. When scheduler is selecting
a cpu for a task, it will check if cpu util is over its capacity, what
could happen is that it will skip a cpu even it is idle, so let scheduler
skip util checking if the task perfers idle cpu and the cpu is idle.

Bug: 172781646
Signed-off-by: Rick Yiu <rickyiu@google.com>
Change-Id: I2e97d36b9658947f702a2c15c9c589a94ed29fb3