dma-heap: samsung: do not release unprotected buffer on error
If releasing H/W protection of a buffer fails, the buffer might be
unusable by Linux forever because we do not have a way to determine if
the buffer is accessible in Linux. So, we should mark that buffer as
unusable by holding the allocated buffer.

However, there was one error handling code path where we did not follow
this rule. Let's fix this by not releasing the buffer if unprotection
fails.

Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>
Change-Id: I2a5f2e0e2159e2fb0858aa1ab662decb23478198