clocksource: exynos_mct: Limit usage only to broadcast device on CPU 0
All arm64 SoCs have an architected timer exposed via a per-cpu system
register interface which is typically low-latency and implemented as
part of the CPU.

The 'exynos_mct' driver deliberately sets its clockevents rating field
to override the architected timer, as this was apparently preferred for
an old 32-bit SoC with shared hardware blocks for implementing the
counter.

Unfortunately, the 'exynos_mct' is relatively slow to access, so reverse
the decision made in 6282edb72bed5 ("clocksource/drivers/exynos_mct:
Increase priority over ARM arch timer") and prefer the architected timer
instead. Since the architected timer cannot be used to wake up from
cpuidle states, the MCT on CPU 0 is used as a broadcast device to wake
others from idle.

Bug: 185092876
Fixes: 6282edb72bed5 ("clocksource/drivers/exynos_mct: Increase priority over ARM arch timer")
Change-Id: I4e838d14b9204eed095ab260fa41e5048cf7c9b1
Signed-off-by: Will Deacon <willdeacon@google.com>