build_slider: Add checking for compatibility with prebuilts
Currently, the prebuilts are updated to match the latest merge point
from AOSP, with the default behavior of build_slider.sh being to use
these prebuilts instead of building the aosp/ directory to cut down on
overall build time. However, if additional commits or uncommitted
changes which deviate from the prebuilts are present, the user should be
made aware and asked to explicitly choose whether or not to use the
prebuilts.

Bug: 184667897
Signed-off-by: J. Avila <elavila@google.com>
Change-Id: I16cbf2d0b479e10c569d08d884867bf24e1cc0ba