arm64/dts: reduce pktproc buffer size
The Pktproc buffer and the DIT buffer have the same size.
If the buffer size is too large,
all buffers cannot be emptied and napi continues to operate.
In this case, when the cluster for IRQ_DIT is changed,
NET_RX may be prevented from being changed to the corresponding cluster.
This is to ensure cluster change while napi stops
by reducing the buffer size.

DL pktproc: 96MB -> 28MB
UL pktproc: 32MB -> 4MB

Signed-off-by: Namkyu Kim <namkyu78.kim@samsung.com>
Change-Id: I511d2adf6264a35368717d2bce3c53b017e9f74f