google/debug: SJTAG kernel-to-kernel interface
With SJTAG authentication implemented, it is now possible to manage
access to privileged resources more prudently, based on whether SJTAG
is being enforced or not AND whether in the enforcing case SJTAG
has been successfully authenticated or not. This CL implements the
above idea. (Note however that issuing this check in the C2 PM
notifier functions turned out to be unnecessary.)

Bug: 184766276
Signed-off-by: Peter Csaszar <pcsaszar@google.com>
Change-Id: I8cdd4c85095fe982fcb2d726c7506fd5d306e702