download_gki: Add support for downloading from gsi-sc-dev
The next phase of GKI for P21 is to switch to directly using boot image
binaries from the GKI release. Modify download_gki to grab images from
the GKI release, which includes both full boot images as well as the
various kernel build outputs necessary for mixed build support.

Note that, for now, we need to use fetch_artifact because
download_from_ci doesn't support downloading from our internal builds.
As a result, this will most likely not work with WFH machines; it will,
however, work on any machines located in office or cloudtop instances.
This solution is intended to be temporary while we work on one which
doesn't have these limitations.

Bug: 191764119
Signed-off-by: J. Avila <elavila@google.com>
Change-Id: I8ff957253475d781ff4af35f5e4f7f2547af83e0