gs101: scsi: ufs: avoid redundant key reprogramming
On resume, the KDN keyslots are being reprogrammed twice -- once in bulk
by gsa_kdn_restore_keys(), and once individually by
blk_ksm_reprogram_all_keys().  This causes unnecessary resume latency.

Fix this by doing the bulk restore only, as the bulk restore is faster
than the individual restores.

Note that the bulk restore seemed to be in the wrong place, as it ran
during resume rather than after UFS reset.  (Currently resume seems to
imply reset on this hardware, but the converse doesn't appear to be
true, as resets can happen in other cases such as error handling.)  So,
to avoid a regression where we might have been relying on the individual
restores in some edge cases, move the bulk restore to the place that the
individual restores previously were.  This required using a vendor hook.

Bug: 181905172
Change-Id: I77e4af579b958527f1a9c0ff095ec948a65bcce2
Signed-off-by: Eric Biggers <ebiggers@google.com>