drivers: thermal: support HW/SW shutdown
Due to HW limitation, NTC engine is not reset after reboot, so sometimes
we may get bogus value which translated to very high temperature. This
patch add a init worker to wait for sensor to be ready. This is needed
for enabling kernel backup thermal shutdown

The patch all added setting hot trip during probe which is needed for
enabling HW backup thermal shutdown.

Bug: 194683509
Test: build and boot
Signed-off-by: Wei Wang <wvw@google.com>
Change-Id: I7f56323d21577f4451095e8a97671b3ff671b1df