build.config: Build cloudripper out-of-tree modules in parallel
Take advantage of GNU Make parallelization to significantly speed up
building of external modules.

To build out-of-tree modules, build.sh uses a for loop to iterate over
the list of out-of-tree modules and invokes "make" and "make
modules_install" during each iteration. This type of serialization
prevents build machines with multiple CPU cores to be used to their full
potential. With 22 such out-of-tree modules, this results in a
significant slow down.

To speed up the build, let's take advantage of GNU Make parallelization
by using a makefile instead of a bash script to drive the build process
for our 22 out-of-tree modules. Make will naturally launch various
submakes in parallel.

As a second optimization, we avoid unnecessarily running depmod after
installing each module, because it is sufficient to call depmod once
after all modules have been installed.

On a workstation with an Intel(R) Xeon(R) Gold 6154 CPU @ 3.00GHz and 72
CPU threads, these optimizations saved 117 seconds in build time.

Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: If97226ee8267494387c03b3320ee8b9ad2f87c6e