build_slider: Remove obsolete merge point check
Before the KMI freeze, there was a concern that developers would
encounter boot or compile time issues when trying to check out their
aosp/ directory beyond the current merge point. This is less of an issue
now, and this check has been confusing for some, so we can go ahead and
remove it.

Signed-off-by: J. Avila <elavila@google.com>
Change-Id: I83ee7d43aebfdd97cb8dcb9c11c665ff367b8b79