driver: devfreq: google: Skip load calculation after resetting the load
In calculate_freq(), when there is no event read from the PPC counter,
it will skip the load calculation. In this case, it would vote the
current governor_freq if "total" is non-zero in the tracked load. This
is to prevent calculate_freq() being called with short duration in
between so the PPC event could be 0. At the same time, load tracking
would reset the load everytime the frequency is changed since the bus
condition would vary with different frequencies. However, right now in
calculate_freq() it will skip resetting the load if there's no event
read from the PPC counter uncontionally. As a result, even with no event
from the PPC counter, it could keep voting the current governor_freq
and cause the frequency vote stay high in some cases. This change makes
sure it always resets the load when there is a frequency change.

Also make sure we always update the PPC counter value even the reject
count (pmcnt1) is zero.

Bug: 201692393
Bug: 203462054
Test: Boot & collect traces
Signed-off-by: Jia-yi Chen <jychen@google.com>
Change-Id: Ieb4d35ae005d31feefe172463f4b2b9f7f750b22